--- com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/dto/HubDto.java
@@ -1,4 +1,0 @@
-package com.devsquad10.hub.application.dto;
-
-public class HubDto {
-}
