+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/dto/enums/HubSortOption.java
@@ -1,4 +1,4 @@
-package com.devsquad10.hub.presentation.enums;
+package com.devsquad10.hub.application.dto.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
