+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/dto/res/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.devsquad10.hub.presentation.res;
+package com.devsquad10.hub.application.dto.res;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
