+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/dto/res/HubUpdateResponseDto.java
@@ -1,4 +1,4 @@
-package com.devsquad10.hub.presentation.res;
+package com.devsquad10.hub.application.dto.res;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -11,14 +11,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-/**
- * 허브 목록의 개별 DTO
- */
 @Getter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PagedHubItemResponseDto implements Serializable {
+@Builder
+public class HubUpdateResponseDto implements Serializable {
+
+	// TODO: dto 개선
 
 	private UUID id;
 	private String name;
@@ -26,17 +25,25 @@ public class PagedHubItemResponseDto implements Serializable {
 	private Double latitude;
 	private Double longitude;
 	private LocalDateTime createdAt;
+	private UUID createdBy;
 	private LocalDateTime updatedAt;
+	private UUID updatedBy;
+	private LocalDateTime deletedAt;
+	private UUID deletedBy;
 
-	public static PagedHubItemResponseDto toResponseDto(Hub hub) {
-		return PagedHubItemResponseDto.builder()
+	public static HubUpdateResponseDto toResponseDto(Hub hub) {
+		return HubUpdateResponseDto.builder()
 			.id(hub.getId())
 			.name(hub.getName())
 			.address(hub.getAddress())
 			.latitude(hub.getLatitude())
 			.longitude(hub.getLongitude())
 			.createdAt(hub.getCreatedAt())
+			.createdBy(hub.getCreatedBy())
 			.updatedAt(hub.getUpdatedAt())
+			.updatedBy(hub.getUpdatedBy())
+			.deletedAt(hub.getDeletedAt())
+			.deletedBy(hub.getDeletedBy())
 			.build();
 	}
 }
