+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/dto/res/PagedHubItemResponseDto.java
@@ -1,4 +1,4 @@
-package com.devsquad10.hub.presentation.res;
+package com.devsquad10.hub.application.dto.res;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -11,13 +11,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * 허브 목록의 개별 DTO
+ */
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class HubUpdateResponseDto implements Serializable {
-
-	// TODO: dto 개선
+public class PagedHubItemResponseDto implements Serializable {
 
 	private UUID id;
 	private String name;
@@ -25,25 +26,17 @@ public class HubUpdateResponseDto implements Serializable {
 	private Double latitude;
 	private Double longitude;
 	private LocalDateTime createdAt;
-	private UUID createdBy;
 	private LocalDateTime updatedAt;
-	private UUID updatedBy;
-	private LocalDateTime deletedAt;
-	private UUID deletedBy;
 
-	public static HubUpdateResponseDto toResponseDto(Hub hub) {
-		return HubUpdateResponseDto.builder()
+	public static PagedHubItemResponseDto toResponseDto(Hub hub) {
+		return PagedHubItemResponseDto.builder()
 			.id(hub.getId())
 			.name(hub.getName())
 			.address(hub.getAddress())
 			.latitude(hub.getLatitude())
 			.longitude(hub.getLongitude())
 			.createdAt(hub.getCreatedAt())
-			.createdBy(hub.getCreatedBy())
 			.updatedAt(hub.getUpdatedAt())
-			.updatedBy(hub.getUpdatedBy())
-			.deletedAt(hub.getDeletedAt())
-			.deletedBy(hub.getDeletedBy())
 			.build();
 	}
 }
