+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/application/service/HubService.java
@@ -10,17 +10,17 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.devsquad10.hub.application.dto.req.HubCreateRequestDto;
+import com.devsquad10.hub.application.dto.req.HubSearchRequestDto;
+import com.devsquad10.hub.application.dto.req.HubUpdateRequestDto;
+import com.devsquad10.hub.application.dto.res.HubCreateResponseDto;
+import com.devsquad10.hub.application.dto.res.HubGetOneResponseDto;
+import com.devsquad10.hub.application.dto.res.HubUpdateResponseDto;
+import com.devsquad10.hub.application.dto.res.PagedHubItemResponseDto;
+import com.devsquad10.hub.application.dto.res.PagedHubResponseDto;
 import com.devsquad10.hub.application.exception.HubNotFoundException;
 import com.devsquad10.hub.domain.model.Hub;
 import com.devsquad10.hub.domain.repository.HubRepository;
-import com.devsquad10.hub.presentation.req.HubCreateRequestDto;
-import com.devsquad10.hub.presentation.req.HubSearchRequestDto;
-import com.devsquad10.hub.presentation.req.HubUpdateRequestDto;
-import com.devsquad10.hub.presentation.res.HubCreateResponseDto;
-import com.devsquad10.hub.presentation.res.HubGetOneResponseDto;
-import com.devsquad10.hub.presentation.res.HubUpdateResponseDto;
-import com.devsquad10.hub.presentation.res.PagedHubItemResponseDto;
-import com.devsquad10.hub.presentation.res.PagedHubResponseDto;
 
 import lombok.RequiredArgsConstructor;
