+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/repository/HubRepository.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 
 import org.springframework.data.domain.Page;
 
+import com.devsquad10.hub.application.dto.req.HubSearchRequestDto;
 import com.devsquad10.hub.domain.model.Hub;
-import com.devsquad10.hub.presentation.req.HubSearchRequestDto;
 
 public interface HubRepository {
 	Hub save(Hub hub);
