+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/repository/HubRepositoryImpl.java
@@ -6,9 +6,9 @@ import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Repository;
 
+import com.devsquad10.hub.application.dto.req.HubSearchRequestDto;
 import com.devsquad10.hub.domain.model.Hub;
 import com.devsquad10.hub.infrastructure.repository.JpaHubRepository;
-import com.devsquad10.hub.presentation.req.HubSearchRequestDto;
 
 import lombok.RequiredArgsConstructor;
